feat: add some data validations and a middleware to handleErrors

+++ src/modules/users/useCases/responsible/deleteResponsible/DeleteResponsibleUseCase.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from "tsyringe";
 
+import { AppError } from "../../../../../errors/AppError";
 import { IResponsibleRepository } from "../../../repositories/IResponsibleRepository";
 
 @injectable()
@@ -10,6 +11,11 @@ class DeleteResponsibleUseCase {
   ) {}
 
   async execute(id: string): Promise<string> {
+    const responsible = await this.responsibleRepository.getResponsible(id);
+    if (!responsible) {
+      throw new AppError("Responsible doesn't exist!", 404);
+    }
+
     const deletedId = await this.responsibleRepository.delete(id);
     return deletedId;
   }
