+++ src/modules/users/useCases/ticket/deleteTicket/DeleteTicketUseCase.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from "tsyringe";
 
+import { AppError } from "../../../../../errors/AppError";
 import { ITicketsRepository } from "../../../repositories/ITicketsRespository";
 
 @injectable()
@@ -10,6 +11,10 @@ class DeleteTicketUseCase {
   ) {}
 
   async execute(id: string): Promise<string> {
+    const ticket = this.ticketsRepository.getTicket(id);
+    if (!ticket) {
+      throw new AppError("Ticket doesn't exist!", 404);
+    }
     const deletedId = this.ticketsRepository.delete(id);
     return deletedId;
   }
